mysql config and stuff
