+++ config/development.js
@@ -4,6 +4,7 @@ module.exports = {
 		"user": process.env.MYSQL_USER,
 		"password": process.env.MYSQL_PASSWORD,
 		"database": process.env.MYSQL_DATABASE,
+		"port": process.env.MYSQL_PORT,
 		"multipleStatements": true
 	},
 	"server": {
