+++ server/administrate.js
@@ -12,19 +12,7 @@ const http = require( 'http' )
 const PROJECT_ROOT = path.resolve( './' )
 const UPLOAD_DIR = path.join( PROJECT_ROOT, 'public/uploads/' )
 
-const globalConfigs = require ( '../config/' )
-const config = globalConfigs.mysql;
-
-const pool  = mysql.createPool( {
-
-	host: config.host,
-	user: config.user,	
-	password: config.password,
-	database : config.database,
-	multipleStatements: true,
-	charset: 'UTF8MB4_UNICODE_CI'
-
-} )
+const pool  = require('./mysql').pool
 
 const _promiseFetchForm = ( req ) => {
