+++ server/data.js
@@ -5,19 +5,7 @@ const formidable = require( 'formidable' )
 const _ = require( 'underscore' )
 const stringify = require( 'csv-stringify' )
 
-const globalConfigs = require ( '../config/' )
-const config = globalConfigs.mysql;
-
-const pool  = mysql.createPool( {
-
-	host: config.host,
-	user: config.user,	
-	password: config.password,
-	database : config.database,
-	multipleStatements: true,
-	charset: 'UTF8MB4_UNICODE_CI'
-
-} )
+const pool  = require('./mysql').pool
 
 const _promiseFetchForm = ( req ) => {
